% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_geojson_to_wkt.R
\name{conv_geojson_to_wkt}
\alias{conv_geojson_to_wkt}
\title{Convert GeoJSON to WKT}
\usage{
conv_geojson_to_wkt(
  geojson.path = paste0(system.file(package = "eamenaR"),
    "/extdata/Kfar_Hahoresh.crane_afg_nd_pl.geojson"),
  dataDir = system.file("extdata", package = "iconr"),
  out.dir = "_out",
  verbose = TRUE
)
}
\arguments{
\item{geojson.path}{The path of the GeoJSON file. By default 'caravanserail_polygon.geojson'.}

\item{dataDir}{Path of the folder storing folders of all decorations. Each of these folders as a site name (eg, Ain Ghazal) and contains at least one shapefile (.shp and .dbf and .shx) and one image (.jpg or .tif or .png, etc.). The shapefile is named conventionally with the name of the site, a dot,the name of the decoration, "nd_pl" for nodes POLYGONS (eg, Ain Ghazal.stat_2_nd_pl.shp). The image is named conventionally with the name of the site, a dot, and the name of the decoration (eg, Ain Ghazal.stat_2.tif).}

\item{out.dir}{Path of the output folder. By default "_out/" in the "dataDir" folder.}

\item{verbose}{if TRUE (by default), verbose.}
}
\value{
Create the 'nodes.csv' file into the out folder, return the complete path of the 'nodes.csv' file.
}
\description{
Convert an EAMENA GeoJSON file to a format adapted to the iconr shape analysis.
}
\examples{

Kfar_Hahoresh.crane_afg_nd_pl_polygon.wkt <- conv_geojson_to_wkt()


}
