% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_pg_to_iconr.R
\name{conv_pg_to_iconr}
\alias{conv_pg_to_iconr}
\title{Convert Pg geometries to 'iconr' geometries}
\usage{
conv_pg_to_iconr(
  dataDir = tempdir(),
  Pg.param = NA,
  sqll.obj = NA,
  sqll.ug.pts = NA,
  sqll.edges = NA
)
}
\arguments{
\item{dataDir}{working directory: sites' folders (copied data) and 'out' folder (outputs)}

\item{Pg.param}{list of arguments to connect the PostgreSQL database.
Like: list(driver, name_of_db, host, port, user, password). By default NA}

\item{sqll.obj}{SQL on objects to get images of decorations. By default NA}

\item{sqll.ug.pts}{SQL on GUs with geometries of type POINTS to get shapes. By default NA}

\item{sqll.edges}{SQL on edges}
}
\value{
Decoration's images and dataframe
}
\description{
Convert the graphical units (GUs) geometries stored in PostgreSQL into iconr format.
}
\examples{

}
