% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_pg_to_shp.R
\name{conv_pg_to_shp}
\alias{conv_pg_to_shp}
\title{Convert Pg geometries to SHP geometries}
\usage{
conv_pg_to_shp(
  dataDir = tempdir(),
  Pg.param = NA,
  sqll.obj = NA,
  sqll.ug.pts = NA,
  sqll.ug.lines = NA,
  sqll.ug.polyg = NA,
  exp.edges = FALSE
)
}
\arguments{
\item{dataDir}{working directory: sites' folders (copied data) and 'out' folder (outputs)}

\item{Pg.param}{list of arguments to connect the PostgreSQL database.
Like: list(driver, name_of_db, host, port, user, password). By default NA}

\item{sqll.obj}{SQL on objects to get images of decorations. By default NA}

\item{sqll.ug.pts}{SQL on GUs with geometries of type POINTS to get shapes. By default NA}

\item{sqll.ug.lines}{SQL on GUs with geometries of type LINES to get shapes. By default NA}

\item{sqll.ug.polyg}{SQL on GUs with geometries of type POLYGONS to get shapes. By default NA}

\item{exp.edges}{Export also edges as shapefiles. By default: FALSE}
}
\value{
Decoration's images and shapefiles
}
\description{
Convert the graphical units (GUs) geometries stored in PostgreSQL into shapefiles (SHP) geometries.
}
\examples{
Pg.param. <- list("PostgreSQL",
                  "postgres",
                  "localhost",
                  5432,
                  "postgres",
                  "postgres")
dataPath <- "D:/decorations/"
dataDir <- paste0(dataPath, "analyse")

### SQL on Pg
## Objects
## sqll.obj. <-  "SELECT
## site, numero, img
## FROM objets
## WHERE (objets.site LIKE 'Ain Ghazal' AND objets.numero LIKE 'stat_2') OR
## (objets.site LIKE 'Ain Ghazal' AND objets.numero LIKE 'stat_5_xd') OR
## (objets.site LIKE 'Qarassa' and objets.numero LIKE 'figurine__wx') OR
## (objets.site LIKE 'Jericho' and objets.numero LIKE 'tete_afe')"

## Polygons
## sqll.ug.polyg. <-  "SELECT
## objets.site,
## objets.numero,
## table_noeuds.id,
## table_noeuds.type,
## table_noeuds.technologie as technlg,
## table_noeuds.incomplet as incmplt,
## ST_AsText(table_noeuds.geom_shape) as wkt FROM objets LEFT JOIN table_noeuds
## ON table_noeuds.site = objets.site AND table_noeuds.decor = objets.numero
## WHERE (objets.site like 'Ain Ghazal' AND objets.numero like 'stat_2')
##  OR (objets.site like 'Ain Ghazal' AND objets.numero like 'stat_5_xd')
##  OR (objets.site like 'Qarassa' AND objets.numero like 'figurine__wx')
##  OR (objets.site like 'Jericho' AND objets.numero like 'tete_afe')"

## Run function
## conv_pg_to_shp(dataDir = dataDir,
##                Pg.param = Pg.param.,
##                sqll.obj = sqll.obj.,
##                sqll.ug.polyg = sqll.ug.polyg.)

}
