% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_pg_to_shp.R
\name{convert_pg_to_shp}
\alias{convert_pg_to_shp}
\alias{conv_pg_to_shp}
\title{Convert Pg geometries to shapefiles geometries}
\usage{
conv_pg_to_shp(
  dataDir = tempdir(),
  Pg.param = NA,
  sqll.obj = NA,
  sqll.ug.pts = NA,
  sqll.ug.lines = NA,
  sqll.ug.polyg = NA,
  exp.edges = FALSE
)
}
\arguments{
\item{dataDir}{working directory: sites' folders (copied data) and 'out' folder (outputs)}

\item{Pg.param}{list of arguments to connect the PostgreSQL database. Like: list(driver, name_of_db, host, port, user, password)}

\item{sqll.obj}{SQL on objects}

\item{sqll.ug.pts}{SQL on GUs with geometries of type POINTS}

\item{sqll.ug.lines}{SQL on GUs with geometries of type LINES}

\item{sqll.ug.polyg}{SQL on GUs with geometries of type POLYGONS}

\item{exp.edges}{Export also edges as shapefiles. By default: FALSE}
}
\value{
Decoration's images and shapefiles
}
\description{
Convert the graphical units (GUs) geometries stored in PostgreSQL into shapefiles geometries.
}
\examples{
convert_pg_to_shp(dataDir = dataDir,

}
