% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_dec.R
\name{list_dec}
\alias{list_dec}
\title{Create Decoration's Graphs and Store them in a List}
\usage{
list_dec(imgs, nodes, edges)
}
\arguments{
\item{imgs}{dataframe of decorations.}

\item{nodes}{dataframe of nodes.}

\item{edges}{dataframe of edges.}
}
\value{
A list of \code{igraph} graphs.
}
\description{
Create undirected graphs for each decoration from \code{nodes}, \code{edges} and \code{imgs} dataframes and store the graphs in a list. The join between these dataframes is done on the two fields \code{site} and \code{decor}. Graph names refer to \code{imgs$idf}.
}
\examples{

# Read imgs, nodes and edges dataframes
imgs <- read.table(system.file("extdata", "imgs.csv", package = "iconr"),
                   sep=";", stringsAsFactors = FALSE)
nodes <- read.table(system.file("extdata", "nodes.csv", package = "iconr"),
                    sep=";", stringsAsFactors = FALSE)
edges <- read.table(system.file("extdata", "edges.csv", package = "iconr"),
                    sep=";", stringsAsFactors = FALSE)
# Create the list of graphs
lgrph <- list_dec(imgs, nodes, edges)

# Get the first graph
g <- lgrph[[1]]
g

# Graph name
g$name

# Graph label
g$lbl

# Graph number of nodes
igraph::gorder(g)

# Graph number of edges
igraph::gsize(g)

}
\seealso{
\code{\link[igraph]{graph_from_data_frame}}
}
