% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_aggregate.R
\name{morph_aggregate}
\alias{morph_aggregate}
\title{Resume geometry classes (POINTS, LINES, POLYGON) of GUs for each object}
\usage{
morph_aggregate(
  nodes = NULL,
  ldist = NULL,
  aggr = c("type", "decoration"),
  out.dir = "_out"
)
}
\arguments{
\item{nodes}{Nodes dataframe coming from the 'conv_shp_to_wkt.R' function}

\item{ldist}{A list of distance matrices coming from 'morph_nds_compar(nodes, focus = "dist")'}

\item{aggr}{Aggregate on GUs' types by decorations ("type") or
aggregate by decorations ("decoration").}

\item{out.dir}{Path of the output folder. By default "_out/" in the "dataDir" folder}
}
\value{
Create plots dependingList of matrices
}
\description{
Aggregate matrices of distances by types calculating mean distances for each single types (eg., two eyes on a single face decoration)
or for the whole decoration (faces, eyes, mouths, ...)
}
\examples{

}
