% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_resume.R
\name{morph_resume}
\alias{morph_resume}
\title{Resume geometry classes (POINTS, LINES, POLYGON) of all graphical units for each different object.}
\usage{
morph_resume(
  dataDir,
  nodes = NA,
  out.dir = "_out",
  imgs.format = c(".jpg", ".png", ".gif", ".tiff", "tif")
)
}
\arguments{
\item{dataDir}{path of the folder storing folders of all decorations}

\item{nodes}{nodes dataframe coming from the 'conv_shp_to_wkt.R' function}

\item{out.dir}{path of the output folder. By default "_out/" in the "dataDir" folder}

\item{imgs.format}{accepted picture formats (see magick image_read)}
}
\description{
Create a contactsheet of decoration with information on graphical units (GUs) geometries:
eg., number of Polygons by types, number of Lines by types, etc.
}
