% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same_elements.R
\name{same_elements}
\alias{same_elements}
\title{Number of Equal Elements Between Each Decoration Pair}
\usage{
same_elements(lgrph, nd.var = "type", focus = "nodes")
}
\arguments{
\item{lgrph}{a list of any number of graphs to be pairwise compared. The list can be typically obtained with the function \code{\link[iconr]{list_dec}}.}

\item{nd.var}{an attribute of the graph vertices containing the node variable (ie, field) on which the comparison will be done. By default \code{nd.var = "type"}.}

\item{focus}{either \code{"nodes"} (default) or \code{"edges"} to select the type of elements to be compared for the count.}
}
\value{
A symmetric matrix with the counts of the pairwise coincidences of nodes or edges. The matrix has as row and column names the names of the corresponding graphs in the input list.
}
\description{
Create the (symmetric) dataframe with the count of \strong{common nodes} or \strong{common edges} (see \code{\link[iconr]{list_compar}} for comparison criteria) for each pair of decorations (graphs) from a list. The diagonal of the symmetric dataframe is filled with counts of nodes/edges for each decoration.
}
\examples{

# read imgs, nodes and edges dataframes
imgs <- read.table(system.file("extdata", "imgs.tsv", package = "iconr"),
                   sep="\t",stringsAsFactors = FALSE)
nodes <- read.table(system.file("extdata", "nodes.tsv", package = "iconr"),
                    sep="\t",stringsAsFactors = FALSE)
edges <- read.table(system.file("extdata", "edges.tsv", package = "iconr"),
                    sep="\t",stringsAsFactors = FALSE)
lgrph <- list_dec(imgs,nodes,edges)

# Counting same nodes
df.same_nodes <- same_elements(lgrph, nd.var = "type",
                               focus = "nodes")
df.same_nodes
## a symmetric matrix of nodes comparisons

# same edges
df.same_edges <- same_elements(lgrph, nd.var = "type",
                               focus = "edges")
df.same_edges
## a symmetric matrix of edges comparisons

}
\seealso{
\code{\link[iconr]{list_dec}}, \code{\link[iconr]{list_compar}}, \code{\link[iconr]{plot_compar}}.
}
