% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_wkt_to_jpg.R
\name{conv_wkt_to_jpg}
\alias{conv_wkt_to_jpg}
\title{Convert WKT geometries to JPG}
\usage{
conv_wkt_to_jpg(nodes, out.dir = "_out")
}
\arguments{
\item{nodes}{nodes dataframe coming from the 'conv_shp_to_wkt.R' function}

\item{out.dir}{path of the output folder. By default "_out/" in the "dataDir" folder}
}
\value{
JPGs files exported into as many folders as different GUs' types (eg., 'bouche', oeil', 'visage', etc.)
The names of the JPGs is the concatenate of their site name, dot, decoration name, dot, GU identifier (eg, "Ain Ghazal.stat_2.1.jpg")
}
\description{
Convert Well-Known Text geometries of graphical units (GUs) to JPG files
 in order to perform contour analysis with the Momocs package
}
\examples{
conv_wkt_to_jpg(nodes = nodes)

## Saving 4.33 x 3.94 in image
## Saving 4.33 x 3.94 in image
## ...
}
