% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_wkt_to_jpg.R
\name{conv_wkt_to_jpg}
\alias{conv_wkt_to_jpg}
\title{Convert WKT geometries to JPG}
\usage{
conv_wkt_to_jpg(nodes, out.dir = "_out")
}
\arguments{
\item{nodes}{nodes dataframe coming from the 'conv_shp_to_wkt.R' function}

\item{out.dir}{path of the output folder. By default "_out/" in the "dataDir" folder}
}
\value{
JPGs files exported into as many folders as different on GUs' type (eg. 'oeil', 'visage', etc.)
}
\description{
Convert Well-Known Text geometries of graphical unites (GUs) to JPG files in order to perform contour analysis with Momocs
}
\examples{
conv_wkt_to_jpg(nodes = nodes)

## Saving 4.33 x 3.94 in image
## Saving 4.33 x 3.94 in image
## ...
}
