% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_nds_compar.R
\name{morph_nds_compar}
\alias{morph_nds_compar}
\title{Morphometrics comparisons between GUs}
\usage{
morph_nds_compar(
  nodes = NA,
  focus = c("panel", "stack", "PCA", "dist"),
  nb.h = 15,
  out.dir = "_out"
)
}
\arguments{
\item{nodes}{Dataframe of nodes}

\item{focus}{Type of analysis: 'panel', 'stack', 'PCA' or 'dist'.
By default c("panel", "stack", "PCA", "dist")}

\item{nb.h}{number of Fourier harmonics, uniquely for PCA. By default = 15}

\item{out.dir}{path of the output folder. By default "_out/" in the "dataDir" folder}
}
\value{
Depending on the focus, return 'panel', 'stack' or 'PCA' plots with their complete path
}
\description{
Momocs package morphometrics comparisons between different graphical units (GUs).
 Read JPG files from each different folder. Useful before grouping (see, 'morph_nds_group' function)
 to determine the number of clusters
}
\examples{
morph_nds_compar(nodes)

## [1] "* read 'oeil' type of UGs"
## Extracting 10.jpg outlines...
## [ 1 / 10 ]  Ain Ghazal.stat_2.1.jpg
## ...
## [ 10 / 10 ]  Qarassa.figurine__wx.14.jpg

}
