% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_resume.R
\name{morph_resume}
\alias{morph_resume}
\title{Resume geometry classes (POINTS, LINES, POLYGON) of GUs for each object}
\usage{
morph_resume(
  dataDir,
  nodes = NA,
  out.dir = "_out",
  imgs.format = c(".jpg", ".png", ".gif", ".tiff", "tif")
)
}
\arguments{
\item{dataDir}{Path of the folder storing folders of all decorations}

\item{nodes}{Nodes dataframe coming from the 'conv_shp_to_wkt.R' function}

\item{out.dir}{Path of the output folder. By default "_out/" in the "dataDir" folder}

\item{imgs.format}{Accepted picture formats (see 'magick' package, 'image_read' function)}
}
\value{
Create a contact sheet of decoration with information on graphical units (GUs) geometries:
eg., number of Polygons by types, number of Lines by types, etc.
}
\description{
Create a contact sheet of decoration with information on graphical units (GUs) geometries:
eg., number of Polygons by types, number of Lines by types, etc.
}
\examples{
> morph_resume(dataDir = dataDir,
+              nodes = nodes)
[1] "Ain Ghazal"
[1] "Jericho"
[1] "Qarassa"
}
